feat: auth (#26)
* feat: add jwt dependencies


* feat: add findByEmail


* feat: new exception types


* refactor: dependency inject services


* feat: add auth service with create method for jwt and basic auth


* chore: use junit5 instead of 4


* test: auth service


* feat: hash password


* fix: messages containing ; had a cut-off body


* refactor: improve failed validation message


* fix: userByEmail returned only deleted users


* feat: add user to auth


* fix: add missing role field to user dto


* feat: add login method for tests


* refactor: test naming with spaces


* feat: parameterized tests, wip users authorization


* feat: parameterized tests, add authorization to users service


---------

+++ src/main/kotlin/services/users/UsersService.kt
@@ -1,12 +1,15 @@
 package at.eventful.messless.services.users
 
 import at.eventful.messless.errors.responses.BadRequest
+import at.eventful.messless.errors.responses.Forbidden
 import at.eventful.messless.errors.responses.NotFound
+import at.eventful.messless.errors.responses.Unauthorized
 import at.eventful.messless.plugins.socket.ServiceMethod
 import at.eventful.messless.plugins.socket.WebSocketService
 import at.eventful.messless.plugins.socket.model.WebSocketResponse
 import at.eventful.messless.repositories.users.commands.UpdateUserCmd
 import at.eventful.messless.schema.dto.UserDto
+import at.eventful.messless.schema.utils.UserRole
 import io.ktor.http.*
 import io.ktor.server.application.*
 import io.ktor.server.plugins.di.*
@@ -16,50 +19,75 @@ import repositories.users.commands.CreateUserCmd
 class UsersService(app: Application) : WebSocketService("users") {
     val usersRepo: UserRepository by app.dependencies
 
+    // TODO: On boot, we need to create an admin user if none exists
     override fun ServiceMethod.create(): WebSocketResponse<UserDto> {
-        val cmd = incoming.receiveBody<CreateUserCmd>()
-
-        try {
-            return WebSocketResponse.from(
-                HttpStatusCode.Created,
-                UserDto.from(usersRepo.addUser(cmd)),
-            )
-        } catch (e: Exception) {
-            if (e.message?.contains("PUBLIC.USERS_EMAIL_UNIQUE") ?: false) {
-                throw BadRequest("A user with this email already exists.")
+        connection.auth.auth?.let {
+            val cmd = incoming.receiveBody<CreateUserCmd>()
+            if (it.user.role != UserRole.Admin && cmd.role.asInt() > it.user.role.asInt()) throw Forbidden("You are not allowed to create a user with this role!")
+            // TODO: add user to company
+            try {
+                return WebSocketResponse.from(
+                    HttpStatusCode.Created,
+                    UserDto.from(usersRepo.addUser(cmd)),
+                )
+            } catch (e: Exception) {
+                if (e.message?.contains("PUBLIC.USERS_EMAIL_UNIQUE") ?: false) {
+                    throw BadRequest("A user with this email already exists.")
+                }
+                throw e
             }
-            throw e
         }
+        throw Unauthorized()
     }
 
     override fun ServiceMethod.find(): WebSocketResponse<List<UserDto>> {
-        return WebSocketResponse.from(
-            HttpStatusCode.OK,
-            usersRepo.allUsers().map(UserDto::from),
-        )
+        connection.auth.auth?.let {
+            return WebSocketResponse.from(
+                HttpStatusCode.OK,
+                usersRepo.allUsers().map(UserDto::from),
+            )
+        }
+        throw Unauthorized()
     }
 
     override fun ServiceMethod.get(id: Int): WebSocketResponse<UserDto> {
-        val user = usersRepo.userById(id) ?: throw NotFound("User with id $id not found")
-        return WebSocketResponse.from(
-            HttpStatusCode.OK,
-            UserDto.from(user),
-        )
+        connection.auth.auth?.let {
+            if (it.user.role != UserRole.Admin && it.user.id != id) throw Forbidden("You are only allowed to query for your own user!")
+
+            val user = usersRepo.userById(id) ?: throw NotFound("User with id $id not found")
+            return WebSocketResponse.from(
+                HttpStatusCode.OK,
+                UserDto.from(user),
+            )
+        }
+        throw Unauthorized()
     }
 
     override fun ServiceMethod.update(id: Int): WebSocketResponse<UserDto> {
-        val updated = usersRepo.updateUser(id, incoming.receiveBody<UpdateUserCmd>())
-            ?: throw NotFound("User with id $id not found")
-        return WebSocketResponse.from(
-            HttpStatusCode.OK,
-            UserDto.from(updated),
-        )
+        connection.auth.auth?.let {
+            if (it.user.role != UserRole.Admin && it.user.id != id) throw Forbidden("You are only allowed to update your own user!")
+
+            val cmd = incoming.receiveBody<UpdateUserCmd>()
+            if (cmd.role.asInt() > it.user.role.asInt()) throw Forbidden("You are not allowed to update to this role!")
+
+            val updated = usersRepo.updateUser(id, cmd) ?: throw NotFound("User with id $id not found")
+            return WebSocketResponse.from(
+                HttpStatusCode.OK,
+                UserDto.from(updated),
+            )
+        }
+        throw Unauthorized()
     }
 
     override fun ServiceMethod.delete(id: Int): WebSocketResponse<Nothing> {
-        usersRepo.removeUser(id) ?: throw NotFound("User with id $id not found")
-        return WebSocketResponse(
-            HttpStatusCode.NoContent
-        )
+        connection.auth.auth?.let {
+            if (it.user.role != UserRole.Admin && it.user.id != id) throw Forbidden("You are only allowed to delete your own user!")
+
+            usersRepo.removeUser(id) ?: throw NotFound("User with id $id not found")
+            return WebSocketResponse(
+                HttpStatusCode.NoContent
+            )
+        }
+        throw Unauthorized()
     }
 }
